fix: reconcile LocalAI conformance with current main
